Added to the vfb the pedal and the steering wheel values. added mayer bug fixes Added a testing log

+++ src/main/java/hu/oe/nik/szfmv/automatedcar/systemcomponents/ACC.java
@@ -72,7 +72,7 @@ public class ACC {
     private int ChangeNewAccSpeed() {
         var currentVelocity = virtualFunctionBus.toPowerTrainPacket.getGasPedalValue();
         if (currentVelocity >= 40 && currentVelocity <= 160) {
-            return currentVelocity;
+            return (int)currentVelocity;
         } else {
             return virtualFunctionBus.inputPacket.getAccSpeed();
         }
