+++ src/main/java/hu/oe/nik/szfmv/automatedcar/virtualfunctionbus/packets/HMIOutputPackets/ReadOnlyToPowerTrainPacket.java
@@ -4,9 +4,9 @@ import hu.oe.nik.szfmv.automatedcar.systemcomponents.Shitfer;
 
 public interface ReadOnlyToPowerTrainPacket {
 
-    int getGasPedalValue();
-    int getBreakPedalValue();
-    int getSteeringWheelValue();
+    double getGasPedalValue();
+    double getBreakPedalValue();
+    double getSteeringWheelValue();
     Shitfer.ShiftPos getShiftChangeRequest();
     int getTempomatValue();
     boolean getTempomatSwitch();
